コンポーネントに切り離す
